Fix misc bugs with formatting

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/auxilary/FileSaver.java
@@ -26,8 +26,8 @@ public class FileSaver {
             this.fileName = _fileName;
             this.filePathname = context.getExternalFilesDir(null).getPath() + "/" + fileName;
         } catch (Throwable ignored) {
-            this.fileName = _fileName;
-            this.filePathname = System.getProperty("java.io.tmpdir");
+            this.fileName = null;
+            this.filePathname = null;
         }
     }
 
@@ -36,6 +36,8 @@ public class FileSaver {
         String thisLine;
         BufferedReader textFile;
         ArrayList<String> lines = new ArrayList<String>();
+
+        if(this.filePathname == null) return lines;
 
         try (BufferedReader br = new BufferedReader(new BufferedReader(new InputStreamReader(new FileInputStream(filePathname))))) {
 
@@ -50,6 +52,8 @@ public class FileSaver {
     }
 
     public String readContent() {
+
+        if(this.filePathname == null) return "";
 
         StringBuilder result = new StringBuilder();
 
@@ -66,11 +70,15 @@ public class FileSaver {
     }
 
     public void deleteFile() {
+        if(this.filePathname == null) return;
+
         File file = new File(filePathname);
         file.delete();
     }
 
     public void overwriteFile(String newContent) {
+        if(this.filePathname == null) return;
+
         try {
             BufferedWriter output = new BufferedWriter(new FileWriter(filePathname));
             output.write(newContent);
@@ -82,6 +90,8 @@ public class FileSaver {
     }
 
     public void appendLine(String line) {
+        if(this.filePathname == null) return;
+
         try {
             BufferedWriter output = new BufferedWriter(new FileWriter(filePathname, true));
             output.newLine();
